+++ js/data.js
@@ -1,0 +1,64 @@
+const AUTOR = [
+  'Вася Пупкин',
+  'Саша Иванов',
+  'Настя Иванова',
+];
+
+const TITLE = [
+  'Отличное предложение',
+  'Хорошие соседи',
+  'Неплохое место',
+];
+
+const TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const CHECKIN = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const CHECKOUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const DESCRIPTION = [
+  'Отличная квартира',
+  'Великолепный вид',
+  'Со всеми удобствами',
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const SIMILAR_OFFER_COUNT = 10;
+
+export {AUTOR};
+export {TITLE};
+export {TYPE};
+export {CHECKIN};
+export {CHECKOUT};
+export {FEATURES};
+export {DESCRIPTION};
+export {PHOTOS};
+export {SIMILAR_OFFER_COUNT};
